requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
openai==1.12.0
google-generativeai==0.3.2
langchain==0.1.20
langchain-openai==0.1.8
langchain-google-genai==0.1.3

# Memory System (Embeddings for semantic search)
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
supabase==2.3.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Web & HTTP
httpx==0.24.1
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
tavily-python>=0.3.0

# Data Processing & Analysis
pandas==2.2.0
numpy==1.26.3
yfinance==0.2.32
ta==0.11.0
sympy==1.12.1

# Visualization
plotly==5.18.0

# Graph & DAG
networkx==3.2.1

# File Processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.11




